api/requirements.txt
azure-functions
#azure-storage-blob
#requests
#openai
#azure-cognitiveservices-speech